\name{print.relative.effect}
\alias{print.relative.effect}
\alias{print.pscore}
\alias{print.stratified.data.frame}   
\alias{print.stratified.pscore}
\alias{print.matched.data.frame}
\alias{print.matched.data.frames}
\alias{print.matched.pscore}
\alias{print.est.data.frame}                   
\alias{print.est.matched.data.frame}
\alias{print.est.matched.pscore}            
\alias{print.est.stratified.data.frame}
\alias{print.est.stratified.pscore}
\alias{print.bal.data.frame}                
\alias{print.bal.matched.data.frame}
\alias{print.bal.matched.pscore}            
\alias{print.bal.stratified.data.frame}
\alias{print.bal.stratified.pscore}
\alias{print.summary.relative.effect}
\alias{print.summary.pscore}
\alias{print.summary.stratified.data.frame}   
\alias{print.summary.stratified.pscore}
\alias{print.summary.matched.data.frame}
\alias{print.summary.matched.data.frames}
\alias{print.summary.matched.pscore}
\alias{print.summary.est.data.frame}                   
\alias{print.summary.est.matched.data.frame}
\alias{print.summary.est.matched.pscore}            
\alias{print.summary.est.stratified.data.frame}
\alias{print.summary.est.stratified.pscore}
\alias{print.summary.bal.data.frame}                
\alias{print.summary.bal.matched.data.frame}
\alias{print.summary.bal.matched.pscore}            
\alias{print.summary.bal.stratified.data.frame}
\alias{print.summary.bal.stratified.pscore}


\title{
  Print function
}

\description{
  to print results.
}

\usage{ \method{print}{relative.effect}(x, ...)
}

\arguments{

  \item{x}{an object to be printed.}

  \item{...}{further arguments passed to.}
}

\details{
  Print function
}

\keyword{
  print
}
