\name{summary.relative.effect}
\alias{summary.relative.effect}
\alias{summary.pscore}
\alias{summary.stratified.data.frame}   
\alias{summary.stratified.pscore}
\alias{summary.matched.data.frame}
\alias{summary.matched.data.frames}
\alias{summary.matched.pscore}
\alias{summary.est.data.frame}                   
\alias{summary.est.matched.data.frame}
\alias{summary.est.matched.pscore}            
\alias{summary.est.stratified.data.frame}
\alias{summary.est.stratified.pscore}
\alias{summary.bal.data.frame}                
\alias{summary.bal.matched.data.frame}
\alias{summary.bal.matched.pscore}            
\alias{summary.bal.stratified.data.frame}
\alias{summary.bal.stratified.pscore}

\title{
  Summary function
}

\description{
  to print summarized results.
}

\usage{ \method{summary}{relative.effect}(object, ...)
}

\arguments{
  
  \item{object}{an object to be printed.}
  \item{...}{further arguments passed to.}
}

\details{
  Summary function
}

\keyword{
  summary
}
